requirements-dev.txt
attrs==22.2.0
cfgv==3.3.1
coverage==7.2.2
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.10.0
identify==2.5.21
iniconfig==2.0.0
model-bakery==1.10.1
nodeenv==1.7.0
packaging==23.0
platformdirs==3.1.1
pluggy==1.0.0
pre-commit==3.2.0
pytest==7.2.2
pytest-django==4.5.2
PyYAML==6.0
requirements-tools==2.1.0
setuptools==67.6.0
tomli==2.0.1
virtualenv==20.21.0
